worker/requirements.txt
praw
redis
rq
cassandra-driver
classtools
